# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "This is a sample swagger for Fiber",
        "title": "My Fucking Swaggah Jagahh.",
        "contact": {},
        "version": "1.0"
    },
    "host": "localhost:3000",
    "basePath": "/",
    "paths": {
...
        "/janus/custSavingInfoJanus/": {
            "post": {
                "description": "/CoreAccounts/API/custSavingInfo",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Janus"
                ],
                "summary": "/CoreAccounts/API/custSavingInfo",
                "parameters": [
                    {
                        "description": "Search",
                        "name": "user",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/route.Acc"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/route.CustSavingInfo"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/route.Errror"
                        }
                    }
                }
            }
        },
        "/janus/customerSavings/": {
            "post": {
                "description": "/CoreAccounts/API/custSavingsList",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Janus"
                ],
                "summary": "/CoreAccounts/API/custSavingsList",
                "parameters": [
                    {
                        "description": "Search",
                        "name": "user",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/route.Use"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/route.CustSavingsListss"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/route.Errror"
                        }
                    }
                }
            }
        },
        "/janus/generateCol/": {
            "post": {
                "description": "CoreAccounts/API/generateColShtperCID",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Janus"
                ],
                "summary": "CoreAccounts/API/generateColShtperCID",
                "parameters": [
                    {
                        "description": "Search",
                        "name": "user",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/route.Use"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/route.GenerateCol"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/route.Errror"
                        }
                    }
                }
            }
        },
        "/janus/loanInfoJanus/": {
            "post": {
                "description": "/CoreAccounts/API/LoanInfo",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Janus"
                ],
                "summary": "/CoreAccounts/API/LoanInfo",
                "parameters": [
                    {
                        "description": "Search",
                        "name": "user",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/route.Acc"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/route.LaonInfo"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/route.Errror"
                        }
                    }
                }
            }
        },
...
        "/janus/multiplePayment/": {
            "post": {
                "description": "/CoreMFS/API/multiplePayment",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Janus"
                ],
                "summary": "/CoreMFS/API/multiplePayment",
                "parameters": [
                    {
                        "description": "Search",
                        "name": "user",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/route.MultiplePaymentRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/route.MultiplePaymentResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/route.Errror"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "route.Acc": {
            "type": "object",
            "properties": {
                "acc": {
                    "type": "string"
                }
            }
        },
        "route.AmortList": {
            "type": "object",
            "properties": {
                "acc": {
                    "type": "string"
                },
                "carVal": {
                    "type": "number"
                },
                "dnum": {
                    "type": "integer"
                },
                "dueDate": {
                    "type": "string"
                },
                "endBal": {
                    "type": "number"
                },
                "endInt": {
                    "type": "number"
                },
                "endOth": {
                    "type": "number"
                },
                "instFlag": {
                    "type": "integer"
                },
                "instPd": {
                    "type": "number"
                },
                "intr": {
                    "type": "number"
                },
                "oth": {
                    "type": "number"
                },
                "penPd": {
                    "type": "number"
                },
                "penalty": {
                    "type": "number"
                },
                "pledgeAmort": {
                    "type": "number"
                },
                "prin": {
                    "type": "number"
                },
                "servFee": {
                    "type": "number"
                },
                "upInt": {
                    "type": "number"
                }
            }
        },
        "route.CustSavingInfo": {
            "type": "object",
            "properties": {
                "acc": {},
                "accDesc": {
                    "type": "string"
                },
                "acctType": {
                    "type": "integer"
                },
                "appType": {
                    "type": "integer"
                },
                "balance": {
                    "type": "number"
                },
                "cid": {
                    "type": "integer"
                },
                "classificationCode": {
                    "type": "integer"
                },
                "classificationType": {
                    "type": "integer"
                },
                "dopen": {
                    "type": "string"
                },
                "status": {
                    "type": "integer"
                }
            }
        },
        "route.CustSavingsListss": {
            "type": "object",
            "properties": {
                "acc": {
                    "type": "string"
                },
                "accDesc": {
                    "type": "string"
                },
                "acctType": {
                    "type": "integer"
                },
                "balance": {
                    "type": "number"
                },
                "cid": {
                    "type": "integer"
                },
                "dopen": {
                    "type": "string"
                },
                "status": {
                    "type": "integer"
                },
                "statusDesc": {
                    "type": "string"
                }
            }
        },
...
        "route.Errror": {
            "type": "object",
            "properties": {
                "errormessage": {
                    "type": "string"
                }
            }
        },
        "route.GenerateCol": {
            "type": "object",
            "properties": {
                "acc": {
                    "type": "string"
                },
                "acctDesc": {
                    "type": "string"
                },
                "accttype": {
                    "type": "integer"
                },
                "amort": {
                    "type": "number"
                },
                "appType": {
                    "type": "integer"
                },
                "area": {
                    "type": "string"
                },
                "areaCode": {
                    "type": "integer"
                },
                "balInt": {
                    "type": "number"
                },
                "balPrin": {
                    "type": "number"
                },
                "centerCode": {
                    "type": "integer"
                },
                "centerName": {
                    "type": "string"
                },
                "classDesc": {
                    "type": "string"
                },
                "classification": {
                    "type": "integer"
                },
                "clientName": {
                    "type": "string"
                },
                "code": {
                    "type": "integer"
                },
                "dateStart": {
                    "type": "string"
                },
                "disbdate": {
                    "type": "string"
                },
                "dueInt": {
                    "type": "number"
                },
                "duePrin": {
                    "type": "number"
                },
                "gives": {
                    "type": "integer"
                },
                "ibalInt": {
                    "type": "number"
                },
                "ibalPrin": {
                    "type": "number"
                },
                "iiid": {
                    "type": "integer"
                },
                "interest": {
                    "type": "number"
                },
                "loanBal": {
                    "type": "number"
                },
                "manCode": {
                    "type": "integer"
                },
                "maturity": {
                    "type": "string"
                },
                "pay": {
                    "type": "number"
                },
                "principal": {
                    "type": "number"
                },
                "saveBal": {
                    "type": "number"
                },
                "staffName": {
                    "type": "string"
                },
                "status": {
                    "type": "integer"
                },
                "statusDesc": {
                    "type": "string"
                },
                "type": {
                    "type": "integer"
                },
                "um": {
                    "type": "string"
                },
                "unit": {
                    "type": "string"
                },
                "unpaidCtr": {
                    "type": "integer"
                },
                "uuid": {},
                "waiveInt": {
                    "type": "number"
                },
                "withdraw": {
                    "type": "number"
                },
                "writeoff": {
                    "type": "integer"
                },
                "writtenoff": {
                    "type": "integer"
                }
            }
        },
...
        "route.LaonInfo": {
            "type": "object",
            "properties": {
                "acc": {
                    "type": "string"
                },
                "accDesc": {
                    "type": "string"
                },
                "acctType": {
                    "type": "integer"
                },
                "amort": {},
                "amortList": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/route.AmortList"
                    }
                },
                "amortcond": {
                    "type": "integer"
                },
                "amortcondvalue": {
                    "type": "number"
                },
                "annumdiv": {
                    "type": "integer"
                },
                "appType": {
                    "type": "integer"
                },
                "approvby": {
                    "type": "string"
                },
                "balance": {
                    "type": "number"
                },
                "charges": {},
                "cid": {
                    "type": "integer"
                },
                "classification_code": {
                    "type": "integer"
                },
                "classification_type": {
                    "type": "integer"
                },
                "conintrate": {
                    "type": "number"
                },
                "cycle": {
                    "type": "integer"
                },
                "disbby": {
                    "type": "string"
                },
                "discounted": {
                    "type": "number"
                },
                "domaturity": {
                    "type": "string"
                },
                "dopen": {
                    "type": "string"
                },
                "frequency": {
                    "type": "integer"
                },
                "fundsource": {
                    "type": "integer"
                },
                "interest": {
                    "type": "number"
                },
                "intr": {
                    "type": "number"
                },
                "isLumpsum": {
                    "type": "integer"
                },
                "lngrpcode": {
                    "type": "integer"
                },
                "loanID": {},
                "netproceed": {
                    "type": "number"
                },
                "oth": {
                    "type": "number"
                },
                "others": {
                    "type": "number"
                },
                "penalty": {
                    "type": "number"
                },
                "prin": {
                    "type": "number"
                },
                "principal": {
                    "type": "number"
                },
                "proff": {
                    "type": "integer"
                },
                "remarks": {},
                "status": {
                    "type": "integer"
                },
                "term": {
                    "type": "integer"
                },
                "waivedint": {
                    "type": "number"
                },
                "weekspaid": {
                    "type": "integer"
                }
            }
        },
...
        "route.MultiplePaymentRequest": {
            "type": "object",
            "properties": {
                "orNumber": {
                    "type": "integer"
                },
                "particulars": {
                    "type": "string"
                },
                "payment": {},
                "prNumber": {
                    "type": "string"
                },
                "remitterCID": {
                    "type": "string"
                },
                "sourceId": {
                    "type": "integer"
                },
                "totalCollection": {
                    "type": "integer"
                },
                "trndate": {
                    "type": "string"
                },
                "username": {
                    "type": "string"
                }
            }
        },
        "route.MultiplePaymentResponse": {
            "type": "object",
            "properties": {
                "message": {
                    "type": "string"
                },
                "retcode": {
                    "type": "integer"
                }
            }
        },
...
        "route.Use": {
            "type": "object",
            "properties": {
                "cid": {
                    "type": "integer",
                    "example": 9
                }
            }
        }
    }
}